Refactor and split public skyproj and private _skyproj.
